backend/pyproject.toml
[project]
name = "invest-simulator"
version = "0.1.0"
description = ""
authors = [
    {name = "kristinabykova",email = "kristina.bykova.03@list.ru"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.135.2,<0.136.0)",
    "uvicorn[standard] (>=0.42.0,<0.43.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "redis (>=7.4.0,<8.0.0)",
    "pyjwt[crypto] (>=2.12.1,<3.0.0)",
    "pwdlib (>=0.3.0,<0.4.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "httpx (==0.24.1)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "openai (==1.30.1)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=26.3.1,<27.0.0)"
]

[tool.poetry]
package-mode = false